+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update RenegadeVikings to 1.2.8, but still ignore T7 vikings (Ashlands).
+
 ## [0.6.31] - 2024-07-06
 
 ### Changed